Refactored matches.js

+++ routes/matches.js
@@ -11,45 +11,51 @@ const Match = require("../models/Match.js")
 router.get("/", isLoggedIn, async (req, res) => {
   const id = req.session.userId
 
-  let user
+  let user,
+    usersPromises = undefined
   try {
     user = await User.findById(id)
   } catch (error) {
     return res.status(500).send("Internal Server Error")
   }
 
-  const usersPromises = user.matches.map(({ userId, matchId }) => {
-    return new Promise((resolve, reject) => {
-      void (async () => {
-        try {
-          const _user = await User.findById(userId)
+  if (user.matches.length > 0) {
+    usersPromises = user.matches.map(({ userId, matchId }) => {
+      return new Promise((resolve, reject) => {
+        void (async () => {
+          try {
+            const _user = await User.findById(userId)
 
-          // If the matched user doesn't exist, remove from match
-          if (!_user) {
-            try {
-              await Match.findOneAndDelete({ _id: matchId })
-            } catch (error) {
-              console.log(
-                "Something went wrong while trying to delete match",
-                matchId
+            // If the matched user doesn't exist, remove from match
+            if (!_user) {
+              try {
+                await Match.findOneAndDelete({ _id: matchId })
+              } catch (error) {
+                console.log(
+                  "Something went wrong while trying to delete match",
+                  matchId
+                )
+              }
+
+              user.matches = user.matches.filter(
+                (matchObj) => matchObj.userId != userId
               )
+              user.save()
+
+              resolve(null)
             }
 
-            user.matches = user.matches.filter(
-              (matchObj) => matchObj.userId != userId
-            )
-            user.save()
-
-            resolve(null)
+            resolve({ ...(await User.findById(userId))._doc, matchId })
+          } catch (error) {
+            reject(error)
           }
-
-          resolve({ ...(await User.findById(userId))._doc, matchId })
-        } catch (error) {
-          reject(error)
-        }
-      })()
+        })()
+      })
     })
-  })
+  }
+
+  if (usersPromises == undefined)
+    return res.status(200).render("matches", { matches: [] })
 
   try {
     const users = await Promise.all(usersPromises)
@@ -57,7 +63,7 @@ router.get("/", isLoggedIn, async (req, res) => {
 
     const renderData = { matches: filteredUsers || [] }
 
-    res.render("matches", renderData)
+    res.status(200).render("matches", renderData)
   } catch (error) {
     res.status(500).send("Internal Server Error", error)
   }
